blkdev_image: add check for copy failure
A common issue is filesystem copy failures due to the filesystem growing
in size and exceeding the size passed to the function. Note that the
intention is to keep the filesystem as small as possible to minimize the
flash write time when flashing. Note that the filesystem and partition will
be grown to the extents of the device on first boot.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>